rename & pindah file trait sesuai review
